requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.23.2
gradio==3.43.0

langchain>=0.3.0,<0.4.0
langchain-openai>=0.3.34
llama-index>=0.11.0

sqlalchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==1.0.0

pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.0.0

pyspark==3.5.3

requests==2.31.0
pandas==2.1.0

dbt-core==1.10.13
dbt-postgres==1.9.1
PyYAML>=6.0
jinja2>=3.1.3,<4.0

pytest==8.4.2
